[eas-cli] release rollout into developer preview
